Fix infinite scroll and move it to custom hook.
